Move (User)Settings tests into new structure
